+++ src/decision_node/src/strategy_node.cpp
@@ -18,6 +18,7 @@
 #include <geometry_msgs/PointStamped.h>
 #include <std_msgs/Bool.h>
 #include <std_msgs/Int32.h>
+#include <std_msgs/UInt8.h>
 
 #include <algorithm>
 #include <cctype>
@@ -666,8 +667,8 @@ int main(int argc, char** argv)
 
   ros::Publisher goal_pub = nh.advertise<geometry_msgs::PointStamped>("clicked_point", 1);
   ros::Publisher motion_pub = nh.advertise<std_msgs::UInt8>("motion", 1);
-  ros::Publisher recover_pub = nh.advertise<std_msgs::Int32>("recover", 1);
-  ros::Publisher bullet_up_pub = nh.advertise<std_msgs::Int32>("bullet_up", 1);
+  ros::Publisher recover_pub = nh.advertise<std_msgs::UInt8>("recover", 1);
+  ros::Publisher bullet_up_pub = nh.advertise<std_msgs::UInt8>("bullet_up", 1);
 
   int tick_hz = kDefaultTickHz;
   pnh.param("tick_hz", tick_hz, tick_hz);
